feat: break out setup seed utils
